usm: tests: classification: Move kafka classification test to USM linux file
